move sample to fiber-types
