+++ crates/fiber-lib/src/fiber/payment.rs
@@ -24,18 +24,16 @@ use crate::fiber::{
 use crate::invoice::{CkbInvoice, InvoiceStore, PreimageStore};
 use crate::Error;
 use crate::{debug_event, now_timestamp_as_millis_u64};
+use ckb_hash::blake2b_256;
+use ckb_types::packed::{OutPoint, Script};
+pub use fiber_types::PaymentSession;
+pub use fiber_types::SendPaymentData;
 use fiber_types::{
     Attempt, BasicMppPaymentData, EntityHex, Hash256, HashAlgorithm, HopHint, PaymentCustomRecords,
     PaymentHopData, PaymentStatus, PeeledPaymentOnionPacket, Privkey, Pubkey, RemoveTlcReason,
     RouterHop, TlcErr, TlcErrData, TlcErrorCode, TrampolineContext, DEFAULT_MAX_PARTS,
     DEFAULT_PAYMENT_MPP_ATTEMPT_TRY_LIMIT, USER_CUSTOM_RECORDS_MAX_INDEX,
 };
-// Re-export SendPaymentData and PaymentSession so that downstream code can still import them
-// from crate::fiber::payment.
-use ckb_hash::blake2b_256;
-use ckb_types::packed::{OutPoint, Script};
-pub use fiber_types::PaymentSession;
-pub use fiber_types::SendPaymentData;
 use ractor::{call_t, Actor, ActorProcessingErr};
 use ractor::{concurrency::Duration, ActorRef, RpcReplyPort};
 use rand::Rng;
