+++ crates/fiber-types/Cargo.toml
@@ -69,6 +69,7 @@ web-time = { version = "1.1.0", features = ["serde"] }
 default = []
 cch = ["bitcoin", "lightning-invoice"]
 watchtower = []
+sample = []
 
 [package.metadata.cargo-shear]
 ignored = ["getrandom"]
