+++ crates/fiber-types/src/sample/invoice.rs
@@ -1,10 +1,9 @@
 use crate::invoice::{
     Attribute, CkbInvoice, CkbInvoiceStatus, Currency, InvoiceData, InvoiceSignature,
 };
+use crate::schema::{CKB_INVOICE_PREFIX, CKB_INVOICE_STATUS_PREFIX};
+use crate::HashAlgorithm;
 use core::time::Duration;
-use fiber_types::schema::{CKB_INVOICE_PREFIX, CKB_INVOICE_STATUS_PREFIX};
-/// StoreSample implementation for `CkbInvoice` and `CkbInvoiceStatus`.
-use fiber_types::HashAlgorithm;
 
 use super::{deterministic_hash256, deterministic_recoverable_signature, StoreSample};
